@@ -11,7 +11,8 @@ description: Flutter version of Space Mapper with 2020 upgrades
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 3.0.2+18
+version: 3.1.0+19
+publish_to: none
 
 environment:
   sdk: ">=2.12.0 <3.0.0"
@@ -19,25 +20,30 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  flutter_localizations: 
+    sdk: flutter
   flutter_background_geolocation: ^4.3.0
   shared_preferences: ^2.0.8
-  background_fetch: ^1.0.1
-  url_launcher: ^6.0.12
+  background_fetch: ^1.0.1  
   http: ^0.13.4
   flutter_map: ^0.14.0
-  # flutter_webview_plugin: ^0.4.0
-  webview_flutter: ^2.1.1
+  webview_flutter: ^2.3.1
   sqflite: ^2.0.0+4
   path_provider: ^2.0.5
   share: ^2.0.4
   uuid: 3.0.5
   flutter_form_builder: ^6.1.0+1
   latlong2: ^0.8.1
-
+  url_launcher: ^6.0.12
+  mailto: ^2.0.0
+  geocoding: ^2.0.1
+  faker: ^2.0.0
+  
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.3
-
+  flutter_vector_icons: ^1.0.0
+  
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -56,9 +62,9 @@ flutter:
   # the material Icons class.
   uses-material-design: true
   # To add assets to your application, add an assets section, like this:
-  # assets:
-  #  - images/a_dot_burr.jpeg
-  #  - images/a_dot_ham.jpeg
+  assets:
+   - lang/en.json
+   - lang/es.json
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
