@@ -1,50 +1,103 @@
+import '../app_localizations.dart';
+import '../models/custom_locations.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_background_geolocation/flutter_background_geolocation.dart'
-    as bg;
 
-class STOListView extends StatelessWidget {
+class STOListView extends StatefulWidget {
+  const STOListView({Key? key}) : super(key: key);
+
+  @override
+  _STOListViewState createState() => _STOListViewState();
+}
+
+class _STOListViewState extends State<STOListView> {
+  @override
+  void initState() {
+    super.initState();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: AppBar(title: Text("Location List")),
-        body: FutureBuilder<List>(
-          future: bg.BackgroundGeolocation.locations,
+        appBar: AppBar(
+            title: Text(
+                AppLocalizations.of(context)!.translate("locations_history"))),
+        body: FutureBuilder<List<CustomLocation>>(
+          future: CustomLocationsManager.getLocations(25),
           builder: (context, snapshot) {
-            if (snapshot.hasData) {
-              List? data = snapshot.data;
-              return _jobsListView(data);
-            } else if (snapshot.hasError) {
-              return Text("${snapshot.error}");
+            if (!snapshot.hasData) {
+              return Center(
+                child: CircularProgressIndicator(),
+              );
             }
-            return CircularProgressIndicator();
+            if (snapshot.hasError) {
+              return Container();
+            }
+            return ListView.builder(
+                itemCount: snapshot.data!.length,
+                itemBuilder: (context, index) {
+                  CustomLocation thisLocation = snapshot.data![index];
+                  return Dismissible(
+                    child: _tile(thisLocation, context),
+                    background: Container(
+                      child: Container(
+                        margin: EdgeInsets.only(right: 10.0),
+                        alignment: Alignment.centerRight,
+                        child:
+                            new LayoutBuilder(builder: (context, constraint) {
+                          return new Icon(Icons.delete_forever,
+                              size: constraint.biggest.height * 0.5);
+                        }),
+                      ),
+                      color: Colors.red,
+                    ),
+                    key: new UniqueKey(),
+                    direction: DismissDirection.endToStart,
+                    onDismissed: (direction) async => {
+                      await thisLocation.deleteThisLocation(),
+                      ScaffoldMessenger.of(context).showSnackBar(
+                          new SnackBar(content: new Text("Location removed")))
+                    },
+                  );
+                });
           },
         ));
   }
 
-  ListView _jobsListView(data) {
-    return ListView.builder(
-        itemCount: data.length,
-        itemBuilder: (context, index) {
-          var thisLocation = data[index];
-          return _tile(
-              thisLocation['timestamp'] +
-                  " activity: " +
-                  thisLocation['activity']['type'].toString(),
-              thisLocation['coords'].toString(),
-              Icons.gps_fixed);
-        });
-  }
-
-  ListTile _tile(String title, String subtitle, IconData icon) => ListTile(
-        title: Text(title,
-            style: TextStyle(
-              fontWeight: FontWeight.w500,
-              fontSize: 20,
-            )),
-        subtitle: Text(subtitle),
-        leading: Icon(
-          icon,
-          color: Colors.blue[500],
+  ListTile _tile(CustomLocation thisLocation, BuildContext context) {
+    String title = thisLocation.getLocality() +
+        ", " +
+        thisLocation.getSubAdministrativeArea() +
+        ", " +
+        thisLocation.getISOCountryCode();
+    String subtitle =
+        thisLocation.getTimestamp() + "\n" + thisLocation.getStreet();
+    String text = thisLocation.displayCustomText(10.0, 10.0, context);
+    return ListTile(
+      title: Text(title,
+          style: TextStyle(
+            fontWeight: FontWeight.w500,
+            fontSize: 20,
+          )),
+      subtitle: new RichText(
+        text: new TextSpan(
+          // Note: Styles for TextSpans must be explicitly defined.
+          // Child text spans will inherit styles from parent
+          style: new TextStyle(
+            fontSize: 14.0,
+            color: Colors.black,
+          ),
+          children: <TextSpan>[
+            new TextSpan(
+                text: subtitle,
+                style: new TextStyle(fontWeight: FontWeight.bold)),
+            new TextSpan(text: text),
+          ],
         ),
-      );
+      ),
+      leading: Icon(
+        Icons.gps_fixed,
+        color: Colors.blue[500],
+      ),
+    );
+  }
 }
